api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.0
h11==0.14.0
idna==3.6
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-decouple==3.8
python-dotenv==1.0.1
python-steam-api==1.2.2
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
